requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1

# NLP Stuff
beautifulsoup4>=4.4.1
cssselect>=0.9.2
feedfinder2>=0.0.4
feedparser>=5.2.1
jieba3k>=0.35.1
lxml>=3.6.0
modzy-sdk==0.6.0
nltk>=3.2.1
Pillow>=3.3.0
pythainlp>=1.7.2
python-dateutil>=2.5.3
PyYAML>=3.11
requests>=2.10.0
tinysegmenter==0.3  # TODO(codelucas): Investigate making this >=0.3
tldextract>=2.0.1
newspaper3k
flask_cors>=3.0.9